refactor: feat: 앱 모듈 및 메인 파일 위치 변경하여 NestJS 애플리케이션 구조 설정
