+++ apps/be/app.module.ts
@@ -1,12 +1,10 @@
 import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
 
-import { DatabaseModule } from './infra/database/database.module';
-import { SessionsModule } from './learning/sessions/sessions.module';
-import { SttModule } from './stt/stt.module';
-
-import { QuestionProviderModule } from './modules/question-provider/question-provider.module';
-
+import { DatabaseModule } from './src/infra/database/database.module';
+import { SessionsModule } from './src/learning/sessions/sessions.module';
+import { QuestionProviderModule } from './src/modules/question-provider/question-provider.module';
+import { SttModule } from './src/stt/stt.module';
 
 @Module({
   imports: [
